requirements.txt
pdoc
numpy
pytest
bs4
gitpython
